requirements.txt
streamlit==1.30.0
openai==0.27.0
requests==2.31.0
python-dotenv==0.19.2
SQLAlchemy==1.4.39
aiohttp==3.10.5